requirements.txt
vertexai
google-generativeai
openai
anthropic
cloudpickle
tqdm
boto3
astor
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bokeh
boto==2.49.0
boto3==1.38.25
botocore==1.38.25
Brotli==1.1.0
cachetools==5.5.2
ccxt==4.4.82
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==45.0.2
curl_cffi==0.11.1
cycler==0.12.1
dateparser==1.2.1
decorator==5.2.1
dnspython==2.7.0
dtaidistance==2.3.13
et_xmlfile==2.0.0
eventkit==1.0.3
exchange_calendars==4.10.1
executing==2.2.0
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
ib-insync==0.9.86
idna==3.10
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
multidict==6.4.4
multitasking==0.0.11
narwhals==1.40.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==1.26.4
packaging==24.2
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==15.0.2
pybit==5.10.1
pycares==4.8.0
pycparser==2.22
pycryptodome==3.23.0
pydeck==0.9.1
Pygments==2.19.1
pyluach==2.2.0
pymongo==4.13.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.0
s3transfer==0.13.0
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.3
seaborn==0.13.2
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
streamlit==1.45.1
ta==0.11.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
xyzservices==2025.4.0
yarl==1.20.0
yfinance==0.2.61
gspread==5.12.4
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
binance-futures-connector
tqdm == 4.67.1
gspread_dataframe
openpyxl == 3.1.5
filelock
psutil